threading-tutorial-main/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
greenlet==1.1.0
idna==2.10
importlib-metadata==4.6.0
lxml==4.6.3
multidict==5.1.0
psycopg2-binary==2.9.1
PyYAML==5.4.1
requests==2.25.1
soupsieve==2.2.1
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
zipp==3.5.0
